修复后端编译错误，服务成功启动
🔧 Bug 修复:
- 修复重复的 error 定义 (ErrUserNotFound)
- 添加缺失的 repository 方法 (GetByUserID, CountByCourse)
- 修复类型不匹配问题 (int vs int64, slice 类型)
- 移除未使用的 import 和变量
- 修复 JWT 中间件 import

✅ 服务状态:
- 后端服务：✅ 运行中 (端口 8080)
- PostgreSQL: ✅ 运行中
- Redis: ✅ 运行中
- API 测试：✅ 正常响应
- Swagger 文档：http://localhost:8080/swagger/index.html

📝 新增文件:
- backend/go.sum (依赖锁定)

📊 统计:
- 修改文件：12 个
- 新增代码：~200 行
- 修复 bug: 15+ 个

🚀 下一步:
- 启动前端服务
- 修复沙箱服务构建问题
- 运行完整测试
- 继续 Phase 3 开发

+++ backend/internal/services/course_service.go
@@ -310,7 +310,7 @@ func (s *CourseService) GetCourseStats(ctx context.Context, courseID uuid.UUID) 
 	// Get enrollment count
 	enrollmentCount, err := s.enrollmentRepo.CountByCourse(ctx, courseID)
 	if err != nil {
-		enrollmentCount = course.EnrollmentCount
+		enrollmentCount = int64(course.EnrollmentCount)
 	}
 
 	// Get lesson count
